refactor(loop): split dispatch module by concern into tables/reducer/gates
dispatch.py grandfathered over the module-health cap. Extract by concern
into sibling modules under teatree.loop, preserving behaviour:

- dispatch_tables.py: the static routing maps + DispatchAction data types
- dispatch_reducer.py: pure ScanSignal->action predicates and routers
- dispatch_gates.py: the network/DB/settings seams (live PR state, the
  RedMrFixAttempt ledger, effective-settings resolution)

dispatch.py keeps only the consult order (_dispatch_one) and the per-tick
dispatch() loop, and re-exports the public surface (dispatch,
DispatchAction, ActionPayload, ActionKind) so external call sites are
unchanged. Table constants and helpers shed their leading underscore as
they become a cross-module API within teatree.loop.

Two monkeypatched test seams move with their canonical home:
review_target_is_dead and get_effective_settings are now patched on
dispatch_gates; doc/comment references to the moved symbols are updated.

dispatch.py shrinks 656->155 raw lines (436->124 by the health metric);
each new module is cohesive and under the cap.
